% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGORITHM.R
\name{ALGORITHM_LOUVAIN}
\alias{ALGORITHM_LOUVAIN}
\title{ALGORITHM_LOUVAIN}
\description{
Is a greedy optimization method to extract communities from large networks 
by optimizing the density of edges inside communities to edges outside 
communities.
}
\section{Supported CRITERION}{

\describe{
  \item{MODULARITY}{See \code{\link{CRITERION_MODULARITY}}}
}
}

\section{PARAMETERS}{

This algorithm does not require any parameters.
}

\seealso{
\code{\link{DynComm}}
}
\author{
poltergeist0
}
