% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communities}
\alias{communities}
\alias{DynComm.communities}
\title{communities()}
\usage{
DynComm.communities(dyncomm)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}
}
\value{
a list of all communities
}
\description{
This method returns all communities from the selected post processing
algorithm or the main algorithm, after the last iteration.
}
\examples{
library(DynComm)

Parameters<-matrix(c("e","0.1","w", "FALSE"),ncol=2, byrow=TRUE)
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,Parameters)
dc$addRemoveEdges(
 matrix(
   c(10,20,10,30,20,30,30,60,40,60,40,50,50,70,60,70)
   ,ncol=2,byrow=TRUE)
)

dc$communities()

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
