% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communitiesEdgeCount}
\alias{communitiesEdgeCount}
\alias{DynComm.communitiesEdgeCount}
\title{communitiesEdgeCount()}
\usage{
DynComm.communitiesEdgeCount(dyncomm)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}
}
\value{
the number of community to community edges in the graph
}
\description{
This method returns the number of community to community edges in the graph
from the selected post processing algorithm or the main algorithm, after the 
last iteration.
}
\examples{
library(DynComm)

Parameters<-matrix(c("e","0.1","w", "FALSE"),ncol=2, byrow=TRUE)
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,Parameters)
dc$addRemoveEdges(
 matrix(
   c(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,12,13,1,1,1,2,2,2,18,12,19,20,2,3,11,12,4,9,5,9,22)
      ,ncol=2)
)

dc$communities()
dc$communitiesEdgeCount()

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
