% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynCommMain.R
\docType{data}
\name{CRITERION}
\alias{CRITERION}
\alias{criterion}
\alias{Criterion}
\title{List of available CRITERION (quality measurement functions).}
\format{
A named list with the names of the available CRITERION:
\describe{
 \item{MODULARITY}{
   Newman-Girvan \cr
   See \code{\link{CRITERION_MODULARITY}}
 }
}
}
\usage{
CRITERION
}
\description{
A criterion is used to indicate the proximity of the current grouping 
of vertices (communities) to the optimum one.
}
\details{
Theoretically, the bigger the value returned by the criterion, the closer the
current grouping is to the best possible grouping.

Each CRITERION internally defines two functions. One is used to 
evaluate if moving a vertex from one group (community) to another 
possibly yields a better overall result. The other is used to measure 
the actual overall quality of the entire grouping (current community 
mapping).

Not all criterion might be available for all algorithms. See each algorithms'
help to find which criterion is supported
}
\examples{
CRITERION$MODULARITY

}
\seealso{
\code{\link{DynComm}}
}
\author{
poltergeist0
}
\keyword{datasets}
