% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRITERION.R
\name{CRITERION_MODULARITY}
\alias{CRITERION_MODULARITY}
\title{CRITERION_MODULARITY}
\description{
Newman-Girvan criterion.
}
\section{Supported in ALGORITHM}{

\describe{
  \item{LOUVAIN}{See \code{\link{ALGORITHM_LOUVAIN}}}
}
}

\section{PARAMETERS}{

This criterion does not require any parameters.
}

\seealso{
\code{\link{DynComm}}
}
\author{
poltergeist0
}
