% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynCommPostProcess.R
\docType{data}
\name{POSTPROCESSING}
\alias{POSTPROCESSING}
\title{List of available post processing algorithms.}
\format{
A named list with the names of the available algorithms:
 \describe{
   \item{DENSOPT}{
     Density optimization is an algorithm that provides a community 
	structure based on the increase of the average community density.
     See \code{\link{postProcessDensOpt}}\cr
     @references \insertRef{Sarmento2019Apr}{DynComm}
   }
 }
}
\usage{
POSTPROCESSING
}
\description{
A post processing algorithm is a function that modifies the results 
presented to the user, allowing for limited result manipulation, but does not 
internally modify the results obtained by the algorithm.
}
\details{
As an example, we are only interested in viewing communities larger than some 
value but do not want to actually remove the smaller ones from the graph, 
invalidating possible future processing over them. A post processing 
algorithm can filter the unwanted values from the results or present a more 
compact version of them.
}
\examples{
POSTPROCESSING$DENSOPT

}
\seealso{
\code{\link{DynComm}}
}
\author{
poltergeist0
}
\keyword{datasets}
