% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSTPROCESSING.R
\name{POSTPROCESSING_DENSOPT}
\alias{POSTPROCESSING_DENSOPT}
\title{POSTPROCESSING_DENSOPT}
\description{
Implementation of the density optimization algorithm as a post processing
algorithm.
}
\section{Performance}{

\describe{
  \item{Initialization}{
  Uses a matrix with three columns and a maximum of verticelAll()^2 rows 
  with the edges between vertices and their weight (vertex<->vertex<->weight)
  of the original graph.
  Temporarily stores a copy of the graph to calculate a new community mapping.
  }
  \item{Results}{
  Uses a matrix with two columns and verticesAll() rows with the new community
  mapping (vertex<->community).
  Uses a matrix with three columns and a maximum of 
  communityCount()^2+communityCount() rows with the edges between communities 
  and their weight (community<->community<->weight).
  }
}

#' @section PARAMETERS:
This post processing algorithm does not require any parameters.
}

\seealso{
\code{\link{DynComm}}
}
\author{
poltergeist0
}
